Add files via upload
Road safety is still a major concern because more and more accidents are having a big influence on society and the economy.

Efforts to proactively improve road safety are hampered by the absence of efficient predictive measures to foresee and prevent accidents.

Creating a reliable machine learning model that both predicts and lessens the likelihood of accidents is the difficult part.

The public, law enforcement, and transportation authorities are among the stakeholders who urgently want a focused solution that tackles the unique dynamics of accident-prone locations and significantly lowers the number of traffic accidents.
The absence of efficient predictive methods to foresee and avoid traffic accidents is the issue being addressed. In order to significantly lower the number of traffic accidents, the emphasis is on creating a machine learning model that can forecast accidents with accuracy.